# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentRunReport": {
        "description": "End-of-run report an agent POSTs to ``/api/v1/agents/runs`` (mTLS-authenticated).\n\nThe host is the verified cert fingerprint, never the body (AR-0012). The server re-derives the\naggregate outcome (ok/partial/failed) and totals from ``scopes`` — it never trusts an\nagent-asserted aggregate — so a misreporting agent can only describe its own scopes, not forge\na fleet-wide health verdict.",
        "properties": {
          "agent_version": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Version"
          },
          "finalized": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finalized"
          },
          "finished_at": {
            "format": "date-time",
            "title": "Finished At",
            "type": "string"
          },
          "pushed": {
            "default": 0,
            "minimum": 0.0,
            "title": "Pushed",
            "type": "integer"
          },
          "reported_config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reported Config"
          },
          "scopes": {
            "items": {
              "$ref": "#/components/schemas/AgentRunScopeFrame"
            },
            "maxItems": 4096,
            "title": "Scopes",
            "type": "array"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          }
        },
        "required": [
          "started_at",
          "finished_at"
        ],
        "title": "AgentRunReport",
        "type": "object"
      },
      "AgentRunResult": {
        "description": "Acknowledgement of a recorded agent run (the server-derived outcome).",
        "properties": {
          "host_id": {
            "title": "Host Id",
            "type": "integer"
          },
          "outcome": {
            "title": "Outcome",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "integer"
          }
        },
        "required": [
          "run_id",
          "host_id",
          "outcome"
        ],
        "title": "AgentRunResult",
        "type": "object"
      },
...
      "EntryFrame": {
        "description": "One filesystem entry in a push batch (mirrors backends.FsEntry on the wire).\n\n``partial_hash`` / ``full_hash`` are carried only on a ``mode='fullbit'`` batch; a metadata\nbatch leaves them ``None`` and ingest then leaves the catalogue columns untouched\n(fullbit-dedup data_model_changes). Both are validated as 64-char BLAKE3 hex.",
        "properties": {
          "ctime": {
            "title": "Ctime",
            "type": "number"
          },
          "dev": {
            "default": 0,
            "title": "Dev",
            "type": "integer"
          },
          "flags": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Flags",
            "type": "object"
          },
          "full_hash": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Hash"
          },
          "gid": {
            "title": "Gid",
            "type": "integer"
          },
          "inode": {
            "title": "Inode",
            "type": "integer"
          },
          "is_dir": {
            "title": "Is Dir",
            "type": "boolean"
          },
          "is_symlink": {
            "title": "Is Symlink",
            "type": "boolean"
          },
          "mtime": {
            "title": "Mtime",
            "type": "number"
          },
          "name": {
            "maxLength": 1024,
            "title": "Name",
            "type": "string"
          },
          "partial_hash": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Partial Hash"
          },
          "path": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Path",
            "type": "string"
          },
          "provider_hash": {
            "anyOf": [
              {
                "pattern": "^[A-Za-z0-9+/=_-]{1,128}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Hash"
          },
          "provider_hash_algo": {
            "anyOf": [
              {
                "pattern": "^[a-z0-9._-]{1,32}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Hash Algo"
          },
          "size_logical": {
            "minimum": 0.0,
            "title": "Size Logical",
            "type": "integer"
          },
          "size_on_disk": {
            "minimum": 0.0,
            "title": "Size On Disk",
            "type": "integer"
          },
          "uid": {
            "title": "Uid",
            "type": "integer"
          }
        },
        "required": [
          "path",
          "name",
          "is_dir",
          "is_symlink",
          "size_logical",
          "size_on_disk",
          "mtime",
          "ctime",
          "uid",
          "gid",
          "inode"
        ],
        "title": "EntryFrame",
        "type": "object"
      },
...
      "FinalizeResult": {
        "description": "Server acknowledgement of a post-drain rollup finalize (ADD 09 §8).\n\nThe agent posts this once after its drain so the server recomputes ``subtree_rollup`` (and a\n``size_history`` point) for the calling host's volumes touched since the last finalize.\n``volume_ids`` are exactly the volumes recomputed (empty when nothing changed since last\ntime); ``rollup_rows`` is the total rollup rows written across them. ``dup_groups`` is the\nnumber of report-only duplicate groups rebuilt in the same call after a full-bit ingest (``0``\non a metadata-only deployment, where no full hashes exist — fullbit-dedup, ADD 02 §7.1). The\nhost is the authenticated mTLS identity, never the body (AR-0012).",
        "properties": {
          "dup_groups": {
            "default": 0,
            "title": "Dup Groups",
            "type": "integer"
          },
          "host_id": {
            "title": "Host Id",
            "type": "integer"
          },
          "rollup_rows": {
            "title": "Rollup Rows",
            "type": "integer"
          },
          "volume_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Volume Ids",
            "type": "array"
          }
        },
        "required": [
          "host_id",
          "volume_ids",
          "rollup_rows"
        ],
        "title": "FinalizeResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HostFrame": {
        "description": "Agent self-description; the authoritative identity is the mTLS fingerprint.",
        "properties": {
          "agent_version": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Version"
          },
          "facts": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HostFactsFrame"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "os": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Os"
          }
        },
        "required": [
          "name"
        ],
        "title": "HostFrame",
        "type": "object"
      },
...
      "IngestBatch": {
        "description": "A single resumable, idempotent push from an agent (ADD 02 §7.2).\n\n``removed`` carries the **explicit** deletions the incremental change feed detected, each keyed\non ``(dev, inode)`` (incremental owner ruling: an explicit present/removed_at marker, NOT\nsnapshot-staleness inference). The server marks those ``(host_id, volume_id, dev, inode)`` rows\n``present=False`` and emits a ``DELETE`` change_log row — it never deletes the catalogue row.\nRe-appearing inodes resurrect on the next upsert. A metadata batch may carry removals; a fullbit\nbatch never does (full-bit re-hashes existing files, it does not detect deletions).\n\n``removed_inodes`` is the **DEPRECATED** legacy fallback an old (pre-(dev,inode)) agent sends:\ninode-only removals. It is consulted ONLY when ``removed`` is empty/absent, and an inode-only\nkey can collide across ZFS child datasets on a ``cross_mounts`` volume (the very bug ``removed``\nfixes) — so new agents send ``removed`` instead. Both lists are bounded by the same per-batch\nlimit as ``entries`` so a malformed batch cannot blow up the reconcile step.",
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/EntryFrame"
            },
            "title": "Entries",
            "type": "array"
          },
          "host": {
            "$ref": "#/components/schemas/HostFrame"
          },
          "mode": {
            "pattern": "^(metadata|fullbit)$",
            "title": "Mode",
            "type": "string"
          },
          "removed": {
            "items": {
              "$ref": "#/components/schemas/RemovedEntryIn"
            },
            "title": "Removed",
            "type": "array"
          },
          "removed_inodes": {
            "items": {
              "type": "integer"
            },
            "title": "Removed Inodes",
            "type": "array"
          },
          "snapshot_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Id"
          },
          "volume": {
            "$ref": "#/components/schemas/VolumeFrame"
          }
        },
        "required": [
          "host",
          "volume",
          "mode"
        ],
        "title": "IngestBatch",
        "type": "object"
      },
      "IngestResult": {
        "description": "Server acknowledgement of an accepted batch.\n\n``entries_removed`` is how many of ``removed_inodes`` matched a live catalogue row and were\nflipped to ``present=False`` (incremental reconciliation). ``changes_logged`` is the number\nof ``change_log`` rows the reconciliation emitted (0 when the volume's feed is disabled).",
        "properties": {
          "changes_logged": {
            "default": 0,
            "title": "Changes Logged",
            "type": "integer"
          },
          "entries_received": {
            "title": "Entries Received",
            "type": "integer"
          },
          "entries_rejected": {
            "title": "Entries Rejected",
            "type": "integer"
          },
          "entries_removed": {
            "default": 0,
            "title": "Entries Removed",
            "type": "integer"
          },
          "host_id": {
            "title": "Host Id",
            "type": "integer"
          },
          "snapshot_id": {
            "title": "Snapshot Id",
            "type": "integer"
          },
          "volume_id": {
            "title": "Volume Id",
            "type": "integer"
          }
        },
        "required": [
          "snapshot_id",
          "host_id",
          "volume_id",
          "entries_received",
          "entries_rejected"
        ],
        "title": "IngestResult",
        "type": "object"
      },
...
      "MarkReadRequest": {
        "description": "Mark specific notifications read (empty list is a no-op; use mark-all-read for all).",
        "properties": {
          "ids": {
            "items": {
              "type": "integer"
            },
            "maxItems": 1000,
            "title": "Ids",
            "type": "array"
          }
        },
        "title": "MarkReadRequest",
        "type": "object"
      },
      "MarkReadResult": {
        "description": "How many notifications were newly marked read.",
        "properties": {
          "marked": {
            "title": "Marked",
            "type": "integer"
          }
        },
        "required": [
          "marked"
        ],
        "title": "MarkReadResult",
        "type": "object"
      },
...
      "NotificationListOut": {
        "description": "A page of notifications plus the current unread count (for the bell badge).",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NotificationOut"
            },
            "title": "Items",
            "type": "array"
          },
          "unread_count": {
            "title": "Unread Count",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "unread_count"
        ],
        "title": "NotificationListOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "HTTP API for Fathomline (built on the Fathom engine). The read/query routes are RBAC- and scope-gated; the agent ingest routes authenticate by mTLS client-cert fingerprint behind the proxy; the remediation, preview and deployment surfaces are default-OFF and individually gated. See the project docs for auth and deployment.",
    "summary": "The storage-estate analyzer — read/query surface, agent ingest, and the default-OFF write/deploy surfaces.",
    "title": "Fathomline API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/agents/finalize": {
      "post": {
        "description": "Recompute subtree rollups (and rebuild dedup groups) after the calling host's drain.\n\nRecomputes ``subtree_rollup`` for the host's freshly-ingested volumes (ADD 09 §8) and, when a\nfull-bit pass has landed content hashes, rebuilds the report-only duplicate groups in the same\ntransaction so the ``/duplicates`` view reflects the hashed content (the documented interim for\nthe arq ``dedup`` queue, ADD 02 §7.1). Carries the SAME mTLS + ingest-proxy-secret boundary as\n``/ingest`` (``FingerprintDep``) and is NOT on the human-auth path: the host is the verified\ncert fingerprint, never the body, and finalize only ever recomputes that host's volumes\n(AR-0012). The agent calls this once after its drain; a host with nothing new since its last\nfinalize recomputes nothing and a metadata-only deployment rebuilds zero dup groups.",
        "operationId": "finalize_rollups_api_v1_agents_finalize_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Client-Cert-Fingerprint",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Client-Cert-Fingerprint"
            }
          },
          {
            "in": "header",
            "name": "X-Fathom-Proxy-Secret",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fathom-Proxy-Secret"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinalizeResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Finalize Rollups",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/v1/agents/ingest": {
      "post": {
        "description": "Accept a batch of fs_entry deltas from an authenticated agent.",
        "operationId": "ingest_batch_api_v1_agents_ingest_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Client-Cert-Fingerprint",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Client-Cert-Fingerprint"
            }
          },
          {
            "in": "header",
            "name": "X-Fathom-Proxy-Secret",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fathom-Proxy-Secret"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Batch",
        "tags": [
          "agents"
        ]
      }
    },
...
    "/api/v1/agents/runs": {
      "post": {
        "description": "Record the calling agent's end-of-run outcome (fleet observability).\n\nSame mTLS + ingest-proxy-secret boundary as ``/ingest`` (``FingerprintDep``): the host is the\nverified cert fingerprint, never the body. The server re-derives the aggregate outcome from the\nreported per-scope rows, so a misreporting agent can only describe its own scopes. Best-effort\nfrom the agent's side — a 404-equivalent (unknown fingerprint) is reported as a no-op, never an\nerror, so run-reporting never destabilises an otherwise-good scan.",
        "operationId": "report_run_api_v1_agents_runs_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Client-Cert-Fingerprint",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Client-Cert-Fingerprint"
            }
          },
          {
            "in": "header",
            "name": "X-Fathom-Proxy-Secret",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fathom-Proxy-Secret"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRunReport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Report Run",
        "tags": [
          "agents"
        ]
      }
    },
...
    "/api/v1/notifications": {
      "get": {
        "description": "List notifications for the bell (newest first, scope-filtered) + the unread count.",
        "operationId": "list_notifications_api_v1_notifications_get",
        "parameters": [
          {
            "in": "query",
            "name": "unread_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Unread Only",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Notifications",
        "tags": [
          "notifications"
        ]
      }
    },
...
    "/api/v1/notifications/mark-read": {
      "post": {
        "description": "Mark specific (in-scope, unread) notifications read.",
        "operationId": "mark_read_api_v1_notifications_mark_read_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkReadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkReadResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mark Read",
        "tags": [
          "notifications"
        ]
      }
    },
...
  }
}